ScriptEngine bug Fix and maintenance.
~ Only enabling mono debugging for debug build.
~ Fix the logic to check the field type is a child of EntityClass. Need to only check monotype MONO_TYPE_CLASS.
+ Added Vsync to WindowProps

+++ Hazel/src/Hazel/Scene/SceneSerializer.cpp
@@ -203,7 +203,7 @@ namespace Hazel
 						}
 						case ScriptFieldType::Byte:
 						{
-							out << scriptField.GetValue<uint16_t>();
+							out << scriptField.GetValue<uint8_t>();
 							break;
 						}
 						case ScriptFieldType::UShort:
